+++ NEWS.md
@@ -8,6 +8,8 @@ specified via `s(..., fx = TRUE)`.
 to the Stan backend in `update` and related methods. (#1373, #1378)
 * Allow to retain unused factors levels via `drop_unused_levels = FALSE` 
 in `brm` and related functions. (#1346)
+* Automatically update old default priors based on new input when 
+when updating models via `update.brmsfit`. (#1380)
 
 ### Other Changes